refactor: Auth 도메인 관련 피드백 반영(이메일 유효성 검사)
