+++ src/main/java/subway/AuthenticationPrincipalConfig.java
@@ -1,10 +1,10 @@
 package subway;
 
-import subway.auth.application.AuthService;
-import subway.auth.ui.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import subway.auth.service.AuthService;
+import subway.auth.ui.AuthenticationPrincipalArgumentResolver;
 
 import java.util.List;
