+++ src/main/java/subway/auth/ui/AuthController.java
@@ -5,9 +5,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import subway.auth.application.AuthService;
 import subway.auth.dto.TokenRequest;
 import subway.auth.dto.TokenResponse;
+import subway.auth.service.AuthService;
 import subway.member.application.MemberService;
 
 @RestController
@@ -21,12 +21,16 @@ public class AuthController {
         this.memberService = memberService;
     }
 
-    // TODO: 로그인(토큰 발급) 요청 처리하기
     @PostMapping("/login/token")
     public ResponseEntity tokenLogin(@RequestBody TokenRequest tokenRequest) {
-        if (!memberService.isPossibleLogin(tokenRequest)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if (!tokenRequest.isValidRequest()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("부적절한 요청입니다.");
         }
+
+        if (!memberService.isPossibleLogin(tokenRequest)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("존재하지 않는 유저입니다.");
+        }
+
         TokenResponse tokenResponse = authService.createToken(tokenRequest);
         return ResponseEntity.ok().body(tokenResponse);
     }
