+++ src/main/java/subway/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -5,9 +5,9 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import subway.auth.application.AuthService;
 import subway.auth.domain.AuthenticationPrincipal;
 import subway.auth.infrastructure.AuthorizationExtractor;
+import subway.auth.service.AuthService;
 import subway.member.domain.LoginMember;
 
 import javax.servlet.http.HttpServletRequest;
